+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes to the full browser extension will be documented in this fil
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## Unreleased
+
 ## 0.0.15 - 2025-05-28
 
 ### Added