+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Fixed
 - Do not record empty elements.
+- Correct `waitForMsec` for synthetic scroll to events.
 
 ## 0.1.3 - 2025-06-18