+++ CHANGELOG.md
@@ -3,7 +3,7 @@ All notable changes to the full browser extension will be documented in this fil
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
-## Unreleased
+## 0.1.4 - 2025-06-30
 
 ### Fixed
 - Do not record empty elements.